Quick Bill: An Inventory and Invoice Generation Application.
I have created a project title 'Quick Bill'. This application is made for a Shopkeeper by which they can add their products to their inventory by both manually or by excel file in this application. They can also create their customers in this application and also generate an invoice to them by mentioning the products they have. This a role based application in which an Shopkeeper can also create and verify their executives and which can also perform some of the tasks which is fixed according to their scope of work. As we have also created a dashboard for Shopkeeper by which they can watch their sales, profit and payment generated in number as well as in percentage. The Superadmin can create the account of both Shopkeeper and Executives of particular Shopkeeper by verifying their email by sending an OTP using nodemailer services and app created provided by Gmail. We have used an JSON API to fetch city and state datasets of our country to enhance user experience and make this application as ease of use. This application involves technology like React.js, Node.js, Express.js and MongoDB. We have used mongoose library for making schema or ORM to make this efficient. And for security purpose we have many function to verify datasets on backend and Authentication (token) also plays very important in it.